Create ConcurrentSessionsShow
This is a script to send multiple show commands to Juniper network devices using Netmiko. It will also use concurrent sessions to send the commands to multiple switches at once to speed things up. The show commands need to be added the the config_commands list currently (keep each command in quotes and separate via comma). This could be improved by accepting a list via text file in the future.